ingestion/requirements.txt
fastapi
uvicorn
psycopg2-binary
google-play-scraper
app-store-scraper
langdetect
vaderSentiment
